requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2025.8.3
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.3
pytest-rerunfailures==14.0
requests==2.32.3
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
